requirements.txt
torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0
numpy==1.23.5
opencv-python==4.8.0.76
pillow==8.2.0
scipy==1.13.1
matplotlib==3.9.4
tqdm==4.67.3
tensorboard
tensorboardx==2.6.4
pycocotools==2.0.11
easydict==1.13
matplotlib==3.5.3